move toggle role to children
